+++ doc/change-log.md
@@ -5,6 +5,7 @@
 - `com.lucendar.common.utils.DateTimeUtils.BeijingConv` added `millisToOdt`, `millisToOdtBoxed`
 - Introduce `Ref`
 - Introduce `KV`
+- Introduce `IoUtils`
 
 # v1.0.0
 - Return type of Reply.ofUpdateResult() changed to `Reply[Void]`